avoid php errors
git-svn-id: https://forge.indepnet.net/svn/glpi/trunk@11873 35612d4c-9d10-0410-837c-bf9dd1be60d7

+++ front/document.form.php
@@ -111,7 +111,7 @@ if (isset($_POST["add"])) {
 
 } else if (isset($_POST["deletedocumentitem"])) {
 
-   if (count($_POST["item"])) {
+   if (isset($_POST["item"]) && count($_POST["item"])) {
       foreach ($_POST["item"] as $key => $val) {
          if ($documentitem->can($key, 'w')) {
             $documentitem->delete(array('id' => $key));
